requirements.txt
streamlit==1.28.0
streamlit-folium==0.15.2
folium==0.14.0
geopy==2.4.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
SpeechRecognition==3.10.0
pyaudio==0.2.13
pyttsx3==2.90
pydub==0.25.1
streamlit-option-menu==0.3.6
plotly==5.17.0
pandas==2.1.0